Add directional light back. Tutorial 20 now shows a stationary directional light + moving point light
Discovered that keeping uninitialized uniform structs in the shader is a bad idea. Commenting out the
material.add*LightAttribute() function seems like it would work to disable a light, however the current
shader contains an uninitialized light struct if thats the case. The point light seems to work if you disable
the directional light, but the scene renders blank if you disable the point light but leave the directional enabled.
I suspect I'm running into some undefined behavior in the glsl compiler in this case.